Rename library files according to PlatformIO requirements
